Add decrypter
This one was *really* fun. The trick for the shift of course is the
modulo operator... that way you can keep the shift for each character
within the bounds of the alphabet.

Since I already have the section ID and the encrypted message, I
implemented an alphabet class that returns the shifted character for any
large shift, using the before-mentioned modulo operator. I searched for
pole, and voila!