Write decompression algorithm
We move a pointer through the whole string. If we get a normal
character, we copy it and continue.

If we find the beginning of a parentheses, we check to see if a whole
"command pattern" exists. We get the values of the length and repeat of
the pattern out of it, and then move the pointer past it. We copy the
pattern the proscribed number of times to the new string.