This should be ready for writing the simulation
All the pieces are in play, just got to tell them how to sort.

Basically, this is a "bot distributed" bubble sort, with complex
instructions. I figure I'll just write it so that it'll solve correctly
the first sim, and then I'll put a guard on the `Bot` class so that if
it's holding the correct chips, it'll report it's name.