some slight refactors of a few things.
I've decided to stop simulating the elevator itself... it can't move
without a single chip or rtg anyway, so I may as well just track it's
current position, and move items up or down without it. I don't need to
track it's inventory, because it basically gets added to the current
floor's anyway.