Getting an answer... just not the shortest.
I think I had swapped some xs and ys because I was trying to do the
rotation in representation of the map. Idiot.

After I finally saw the error, and only do the rotation during the
display... we are getting an answer! Now I need to return all the
possible routes and take the shortest.