Check in first version of the one time pad code
I've been struggling with this one off and on for the past day... I'm
not finding all the keys, so that the desired 64th key is too far back
in the list. Ugh.

I'm not sure whether abbbbbc counts as three in a row or not.

It doesn't seem to make a difference. I am trying to go through the list
once and memoizing it to minimize the amount of time I waste generating
MD5 hashes.