Solve Day 15 part 1
After the stupidity of Day 14, I solve Day 15 first try. Bit of a morale
booster. Nothing too clever here, and I'd like to rewrite `solve` to be
a little less hacky. The key part is the `open_at` method, which
correctly handles the rotation of the disk positions by using the ever
handy modulo operator.