Ugly solution to Day 18
Not my finest work, and very slow.

I chose to use hashes instead of arrays, because I could do "fake
bounds" with it. If you go out of bounds on an array, you get a negative
index, and the last element. If you fake an array with a hash, you get
nil.

¯\(°_o)/¯