Write the second half of Day 20
I was pretty sure I could reuse a lot of the first day... I ran into a
small issue... I hadn't read all the requirements. Using a set was
merely to examine and throw away any duplicates (don't know how that
could have snuck in, but it was worth a shot) and so I just leveraged it
to discard the one ip that was too high. :P