Start Day3: finding the circle
The way I see it, I could either try to figure out the math of the
spiral, or model the whole thing.

Even if you modeled the whole thing, the only thing that matters is the
distance from the center (which ring you are on) and how far the current
number is from the midpoint of the ring (the shortest distance to the
center.) Let ring be `r` and distance to closest midpoint on the ring be
`x`, I believe the shortest Manhattan distance will be `r+x`. The
largest possible distance would be on a corner, where `x` is greatest,
and the shortest possible will be on a midpoint, where `x == 0`.

The size of the ring is going to go up by 2s from 1, so I wrote a
function to find out on what ring a number will occur. Let `num` be the
target number, `num` is on the ring (`r`) where `(r-1)^2 > num`.

If you chose to model the whole thing, you could skip the evaluation
step by merely keeping track of the current position as you filled up
"the memory". I think the math way will be easier. Tally ho!