Write Day10 p1
I didn't want to fool with implementing a circular linked list... so I
used the same rotate trick I did in Day 6, except this time I'm adding
in unrotate function so that I don't have to keep track of the indexes.
Much cleaner.

I don't like how the match in the function calls gets a little heavy,
but I'm pretty happy with the result.