First stabs at structure and parsing: FAIL
So not as easy as I was hoping. My naive interpretation works for a few
of the test cases, but ignores one simple fact... Any two directions
separated by a direction makes == the middle seat direction.

- sw + se == s
- sw + n == nw

Also, I guess I don't even have to step through the path if I don't
want to, I could simply hot-load the number of each atom in the input
into the state array, but I'm guessing I'll need the step through for
part II.