Finally finish Day7 P2
Retrospective:
I spent an *ungodly* amount of time trying to homebrew my own tree
structure in elixir, and really suffered as a result. I eventually used
libgraph, which after I grokked it a bit made it relatively easy to
compute values in the tree.

Still not the best code I've ever written. I'm not super proud of it.