Solve p1 and p2
After some tests in IEx, I was pretty convinced that I was done. I need
to rework it dramatically though, it's really slow for p2.

I think that building a full lookup map in the rules is going to help
quite a bit, so going to take a stab at that.