Finish Day14 P2
This was an evil bug. The key part in the instructions:

> To convert to bits, turn each hexadecimal digit to its equivalent
> binary value, high-bit first: 0 becomes 0000, 1 becomes 0001, e becomes
> 1110, f becomes 1111, and so on; a hash that begins with a0c2017... in
> hexadecimal would begin with 10100000110000100000000101110000... in
> binary.

I didn't see it until I wrote a ghetto pretty-print for the grid, and
the case statement started complaining that I didn't have a case for
'nil' values. I double-checked my list to map function... it wasn't
distorting the grid, it was the defrag method that was returning the nil
values in the array... I re-read the above instructions, and it hit
me... the Integer.to_string wasn't padding out the binary number with
zeros. Therefore, p1 was getting the right answer because it's just
counting the 1 values, but the 1 values were not in the right place
because they needed to be shifted over by the right number of zeroes.