Finish Day21 P2
Using @primesaremyfavorite's awesome heurisitic, I was able to finish
P2. The code is very ugly, mostly because I'm transforming a
datastructure up and then back down a series of transformations. I think
if I reconsidered the datastructure from the start, it'd be cleaner, but
maybe I'll come back and refactor it soon.