Starting a solution for Day2
I think I'm going to use maps to represent the keypad, it has the added
advantage of allowing me to have jagged/irregular structures and some
built-in bounds detection.

The only trouble is `get_in` this way is [y, x], so if you have {x, y}
pairs you need a transposed getter function, which is fairly trivial.

iex(12)> get_in(%{0 => %{0 => "a", 1 => "b"}}, [0, 0])
"a"
iex(13)> get_in(%{0 => %{0 => "a", 1 => "b"}}, [0, 3])
nil