Solve part 2016.5.2
I'm not crazy about this solution, there's some *ugly* elixir in here,
but I get results, so that's nice.

I refactored `decode/3` to take another parameter (thus becoming
`decode/4`), this new param is a function that fills the password hash
(which used to be a list) based on the logic for the part of the puzzle.

If I *really* wanted to refactor for speed, I'd probably have to figure out
a way to go through the numbers only once.