Write test solution to 2016.7.1
I sketched out the structure of this for a while, but I kept being hung
up by test failures... I think what happened was that I had a series of
if statements that were occassionally *both* true, so the same `char`
would go down both a `scan` and a `hypernet`. I moved to `cond` just for
show, and it removed the problem and helped me see what I was doing.