REWRITE. I DEMAND A REWRITE.
I went through a couple iterations, I was convinced I needed a struct or
a MapSet at some point, I think I'm just going to give this a whirl. I
think a list is probably a slow implementation for now, and a MapSet
would be *faster*, yet it adds a lot of cruft, and I like how simple
this is so far.