Write first version of 2016.10 in elixir
This has gone through a few writes/rewrites. I've settled on
representing the world as a big ol' map. I'm using some new techniques
using the `update_in`, `put_in`, `pop_in`, to manage state in the nested
map, as well as `Access.key/2` to build the map as it goes.

I'm not super happy with this yet, it's very messy and strange.