Solve 2016.12
This is roughly the pattern I've used for solving these code/state
machine problems... transform the source on load to something easier to
work with, and write a case statement or multiple functions to handle
the assembly style statements.

I think the only new trick was putting the recursion on the end of the
`case` statement instead of on every single branch of the cases
statement. I'm pretty sure I have other puzzles that would benefit from
this.