Rewrite to use `MapSet`
While prepping for our little Advent of Code meeting, I realized that I
could probably speed up my implementation by using a simpler datamodel
for the grid of on/off squares... a set of coords, only storing the
"on" squares. Originally, I was using a set of nested maps, and then I
rewrote to be map of *every* position with either a 1 or a 0 as a
value... this is much simpler, and about 3x faster.