Use more efficient comparison for `judge/2`
I had a hunch that the reason that my solution was so slow was that I
was transforming an integer -> string -> array -> string... not good.

I googled it real quick, and found exactly the solution I was trying to
think, except that I'm terrible at hex/binary maths.

What is happening here (as I understand it) is each number is binary
AND'd with 0xffff (which in binary is of course 1111111111111111). If
the two results match, then they have matching rightmost 16 digits.

Shamelessly stolen from:
https://www.reddit.com/r/adventofcode/comments/7jxkiw/2017_day_15_solutions/dra2hfy/