That's a lot of boilerplate
I'm not crazy about it, but it *is* clear and testable.  I haven't
finished, but it seems to be tracking correctly.

I think the most interesting insight is in `next_orientation/2`. I was
looking for a simple way to determine whether the car is turning left or
right, and the fact that we can determine whether the car is turning
left or right depending on whether it's on the horizontal or vertical
track is interesting.