Finish Day 1
Not super hard, mostly just trying to remember how to elixir.

Lots of nice test cases here... the "crux" of the problem was handling
part 2. I decided to hold onto the history of all of them, and appending
to the head for performance reasons. I had to reverse and `tl/1` the
list to get everything but the first mass to compute the total fuel,
that was the only real hang up.