Solved: 2019.9.1 and 2019.9.2
That feels **really** good.

All the problems I was encountering had to do with how I was handling
the parameter modes. I wasn't reversing them correctly, and for some
reason it didn't matter for my problem input.

After rewriting that, it also became clear that I wasn't taking
parameter modes into account when writing to the tape, so
`Intcode.put/3` had to be written and then implemented everywhere.