Setup final push on 2019.10.2
I think that I'm in a position to be able to sort all the remaining
asteroids and be able to "destroy" them in order.

I'm converting the coordinates to polar coords, and setting them up to
be easily searchable/sortable.

I've got the pseudocode for the next solution, now I just have to write
what I suspect will be either a reduce or recursive function.